chore: 글 목록 모달 컴포넌트화 및 파일 위치 재설정
